backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via
    #   pdfminer-six
    #   requests
click==8.3.1
    # via uvicorn
cryptography==46.0.3
    # via pdfminer-six
distro==1.9.0
    # via openai
faiss-cpu==1.13.2
    # via -r requirements.in
fastapi==0.128.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.12.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==1.2.4
    # via -r requirements.in
langchain-core==1.2.7
    # via
    #   langchain
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-openai==1.1.7
    # via -r requirements.in
langchain-text-splitters==1.1.0
    # via -r requirements.in
langgraph==1.0.6
    # via langchain
langgraph-checkpoint==4.0.0
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.6
    # via langgraph
langgraph-sdk==0.3.3
    # via langgraph
langsmith==0.6.4
    # via langchain-core
numpy==2.4.1
    # via faiss-cpu
openai==2.15.0
    # via langchain-openai
orjson==3.11.5
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.1
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   faiss-cpu
    #   langchain-core
    #   langsmith
pdfminer-six==20251230
    # via pdfplumber
pdfplumber==0.11.9
    # via -r requirements.in
pillow==12.1.0
    # via pdfplumber
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
    #   langchain
    #   langchain-core
    #   langgraph
    #   langsmith
    #   openai
pydantic-core==2.41.5
    # via pydantic
pypdfium2==5.3.0
    # via pdfplumber
python-dotenv==1.2.1
    # via uvicorn
python-multipart==0.0.21
    # via -r requirements.in
pyyaml==6.0.3
    # via
    #   langchain-core
    #   uvicorn
regex==2026.1.15
    # via tiktoken
requests==2.32.5
    # via
    #   -r requirements.in
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via openai
starlette==0.50.0
    # via fastapi
tenacity==9.1.2
    # via langchain-core
tiktoken==0.12.0
    # via langchain-openai
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   fastapi
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
uuid-utils==0.13.0
    # via
    #   langchain-core
    #   langsmith
uvicorn[standard]==0.40.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
xxhash==3.6.0
    # via langgraph
zstandard==0.25.0
    # via langsmith
